[PPC64][libunwind] Fix register r2 not being properly restored
This change makes each unwind step inspect the instruction at the return address and,
if needed, read r2 from its saved location and modify the context appropriately.

One point that (maybe) could be improved is to try to find if the binary being
unwound is ELFv1 or ELFv2 at run time, instead of at compile time, or, find out
the correct save location of r2 in some other way.