[PPC64] Update LocalEntry from assigned symbols
On PowerPC64 ELFv2 ABI, functions may have 2 entry points: global and
local.  The local entry point location of a function is stored in the
st_other field of the symbol, as an offset relative to the global entry
point.

In order to make symbol assignments (e.g. .equ/.set) work properly with
this, PPCTargetELFStreamer already copies the local entry bits from the
source symbol to the destination one, on emitAssignment(). The problem
is that this copy is performed only at the assignment location, where
the source symbol may not yet have processed the .localentry directive,
that sets the local entry. This may cause the destination symbol to end
up with wrong local entry information. Other symbol info is not affected
by this because, in this case, the destination symbol value is actually
a symbol reference.

This change keeps track of these assignments, and update all needed
st_other fields when finish() is called.

(https://reviews.llvm.org/D56586)