[PPC64] Fix ofw_initrd
Before this change, OFW initrd (as md) handling code was simulating an
ofwbus device. But as there isn't really a Device Tree (DT) node
representing OFW initrd (it is specified in 2 properties under
/chosen), its driver was in fact stealing other driver's DT node.
This was noticed after MD_ROOT_MEM became default and QEMU's USB
keyboard stopped working under VNC.

This change consists in simplifying the process of detection and mapping
of initrd memory, turning it into a simple startup step, instead of
trying to simulate a device.