powerpc: EARLY_AP_STARTUP support investigation
Allows that APs can start just after SI_SUB_CONFIGURE sub
system when option EARLY_AP_STARTUP is defined.

It is not possible to start the APs earlier that SI_SUB_CONFIGURE in ppc
(as it is in amd64) because IPI is dependent of external PIC whose
driver is determined at runtime.

More investigation is necessary to see if these changes are enough to
guarantee correct working, for example, a safer approach would be to
start APs after SI_SUB_INT_CONFIG_HOOKS.