Fix mpr driver endianness
This patch fix endiannes issues in the mpr driver when functioning in
a big endian platform. The changes were tested with a SAS9300-8i
card chip LSISAS3008 (pci vendor=0x1000 device=0x0097) on a Talos
and IBM machines with POWER9 processor.
The mpr driver was also added to GENERIC64 conf for ready usage on
future.