powerpc: move pmc_hook out of critical section
Calling pmc_hook inside a critical section may result in a panic.
This happens when the user callchain is fetched, because it uses
pmap_map_user_ptr, that tries to get the (sleepable) pmap lock
when the needed vsid is not found.
Judging by the implementation in other platforms,
intr_irq_handler in kern/subr_intr.c and what pmc_hook do, it
seems safe to move pmc_hook outside the critical section.