Fix OFWFB with Radix MMU
There are some VA ranges that can't be used with Radix (at least not
without modifying how pmap_kenter_attr is implemented in it). One of
those is the VA returned by pmap_early_io_map for OFWFB physical
address.
To fix this issue, we need to either modify radix implementation to
support adding l1/l2 entries when they are not found, or we need to
modify pmap_early_io_map to use a VA range supported by current
radix implementation. This change uses the latter approach, to use
a part of initial VA range for early I/O maps, that can be safely
remapped later.